refactor: review-comment: Make tests adhere to the new structure
