+++ tests/shared/generator_helpers.py
